+++ app/src/main/java/com/example/simbirsoft_android_practice/Publication.kt
@@ -1,7 +1,9 @@
 package com.example.simbirsoft_android_practice
 
+import java.math.BigDecimal
+
 interface Publication {
-    val price: Double
+    val price: BigDecimal
     val wordCount: Int
     fun getType(): String
 }
